feat(console): remove old object explorer
Delete the /objects routes and all object-explorer components, replaced
by the inline data catalog panel. Shared code (NULL_DATABASE_NAME,
SupportedObjectType, objectIcon) is moved to common locations before
deletion:

- NULL_DATABASE_NAME → src/platform/constants.ts
- SupportedObjectType + OBJECT_TYPE_TO_LABEL → src/api/materialize/types.ts
- objectIcon → src/components/objectIcons.tsx
- routerHelpers inlined into src/platform/routeHelpers.ts
