+++ console/src/platform/AuthenticatedRoutes.tsx
@@ -34,8 +34,6 @@ import { EnvironmentNotReadyRoutes } from "~/platform/environment-not-ready/Envi
 import { EnvironmentOverviewRoutes } from "~/platform/environment-overview/EnvironmentOverviewRoutes";
 import IntegrationsRoutes from "~/platform/integrations/IntegrationsRoutes";
 import { MaintainedObjectsRoutes } from "~/platform/maintained-objects/MaintainedObjectsRoutes";
-import { ObjectExplorerDetailRoutes } from "~/platform/object-explorer/ObjectExplorerDetailRoutes";
-import { ObjectExplorerRoutes } from "~/platform/object-explorer/ObjectExplorerRoutes";
 import QueryHistoryRoutes from "~/platform/query-history/QueryHistoryRoutes";
 import RolesRoutes from "~/platform/roles/RolesRoutes";
 import {
@@ -275,17 +273,6 @@ const EnvironmentRoutes = () => {
           </BaseLayout>
         }
       />
-      <Route path="objects">
-        <Route
-          index
-          path="*"
-          element={
-            <BaseLayout sectionNav={<ObjectExplorerRoutes />}>
-              <ObjectExplorerDetailRoutes />
-            </BaseLayout>
-          }
-        />
-      </Route>
       <Route path="clusters">
         <Route
           index
