+++ console/src/platform/clusters/IndexList.tsx
@@ -50,8 +50,8 @@ import { useFlags } from "~/hooks/useFlags";
 import useLocalStorage from "~/hooks/useLocalStorage";
 import { InfoIcon } from "~/icons";
 import { MainContentContainer } from "~/layouts/BaseLayout";
-import { NULL_DATABASE_NAME } from "~/platform/object-explorer/constants";
-import type { SupportedObjectType } from "~/platform/object-explorer/ObjectExplorerNode";
+import { NULL_DATABASE_NAME } from "~/platform/constants";
+import type { SupportedObjectType } from "~/api/materialize/types";
 import { useBuildIndexPath } from "~/platform/routeHelpers";
 import { useAllClusters } from "~/store/allClusters";
 import { useOpenCatalogDetail } from "~/store/catalog";
